requirements.txt
# Self-serve ETL framework - local run
# Python 3.9+

# Data
pandas>=2.0.0
pyarrow>=14.0.0
duckdb>=0.9.0

# Config
PyYAML>=6.0

# dbt
dbt-duckdb>=1.7.0

# optional dev / tests
pytest>=7.0.0
